[TASK] Event selection should respect archive constraint (refs #6)
In order to be able to display only future or only past events, the
archive constraint must be respected in the repository. Furthermore,
the datetime constraint for the calendar plugin should only be applied
if a month and year is requested so it doesn't interfere with an
event list without such restrictions.