Change position for adding of Event Details tab to teaser
Since the bodytext field is not available under External and Internal link types, the event details tab is placed at the back of the list instead of the second tab, like a regular news item. Changing this to teaser will solve the problem, since this field is shown on all types.